pyproject.toml
[project]
name = "rpi-arduino"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5",
]
